requirements.txt
# Core RL and Environment
gymnasium>=0.29.0
numpy>=1.24.0
stable-baselines3>=2.0.0
torch>=2.0.0

# Geospatial and Coordinate Transforms
geopy>=2.3.0
pyproj>=3.5.0

# Visualization and UI
matplotlib>=3.7.0
pyyaml>=6.0
rich>=13.0.0

# Optional: For TensorBoard monitoring
tensorboard>=2.13.0
